Added namespace TIMESERIES to lessen global namespace pollution of all the Timeseries subclasses. Added BCM plasticity, not working yet.

+++ trunk/tau.cpp
@@ -15,7 +15,7 @@ void Tau::init( Configf& configf )
     m.resize(1,temp[0]/deltat);
     max = m[0];
   }
-  else if( temp.size() == nodes ) {
+  else if( temp.size() == uint(nodes) ) {
     for( int i=0; i<nodes; i++ ) {
       if( remainder(temp[0],deltat) >deltat ) {
         std::cerr<<"Value of tau not divisible by Deltat!"<<endl;
